requirements.txt
# Email Processing Tool Dependencies
# Core dependencies
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
schedule==1.2.0
python-dotenv==1.0.0

# Email processing
imapclient==2.3.1
email-validator==2.1.0

# Database
SQLAlchemy==2.0.25
alembic==1.13.1

# Logging and monitoring
structlog==23.2.0
colorlog==6.8.0

# System integration
daemon==1.2
psutil==5.9.7

# Testing (development)
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0 